refactor: Create SquareFactory
